supervisor/backend/requirements.txt
# ============================================
# SUPERVISOR V2.0 - Backend Dependencies
# ============================================
# Python 3.12.7
# Date: 2025-11-11

# ============================================
# Core Framework
# ============================================
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1

# ============================================
# Authentication & Security
# ============================================
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0
PyJWT==2.8.0

# ============================================
# Database
# ============================================
mysqlclient==2.2.4
# Alternative si mysqlclient pose problème: PyMySQL==1.1.0

# ============================================
# Configuration & Environment
# ============================================
python-decouple==3.8
python-dotenv==1.0.0

# ============================================
# Image Processing
# ============================================
Pillow==10.2.0
ExifRead==3.0.0

# ============================================
# Document Generation - Excel
# ============================================
openpyxl==3.1.2
XlsxWriter==3.1.9
pandas==2.1.4

# ============================================
# Document Generation - Word
# ============================================
python-docx==1.1.0
docxtpl==0.16.7

# ============================================
# Document Generation - PowerPoint
# ============================================
python-pptx==0.6.23

# ============================================
# Document Generation - PDF
# ============================================
reportlab==4.0.9
WeasyPrint==60.2
xhtml2pdf==0.2.13

# ============================================
# Geolocation & Mapping
# ============================================
geopy==2.4.1
simplekml==1.3.6
pykml==0.2.0

# ============================================
# Date & Time
# ============================================
python-dateutil==2.8.2
pytz==2023.3.post1

# ============================================
# HTTP Requests & APIs
# ============================================
requests==2.31.0
urllib3==2.1.0

# ============================================
# WhatsApp & Notifications
# ============================================
twilio==8.11.1

# ============================================
# Data Validation & Processing
# ============================================
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
validate-email==1.3

# ============================================
# File Handling
# ============================================
python-magic==0.4.27
# Alternative Windows: python-magic-bin==0.4.14

# ============================================
# OCR & AI (pour analyse de documents)
# ============================================
pytesseract==0.3.10
# Note: Nécessite l'installation de Tesseract OCR séparément

# ============================================
# Utilities
# ============================================
django-filter==23.5
django-extensions==3.2.3

# ============================================
# Monitoring & Logging
# ============================================
django-debug-toolbar==4.2.0
sentry-sdk==1.39.2

# ============================================
# Task Queue (pour tâches asynchrones)
# ============================================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# ============================================
# Code Quality
# ============================================
flake8==7.0.0
black==23.12.1
isort==5.13.2
pylint==3.0.3
pylint-django==2.5.5

# ============================================
# Documentation
# ============================================
drf-yasg==1.21.7
drf-spectacular==0.27.0

# ============================================
# Storage & Cloud (si nécessaire)
# ============================================
# django-storages==1.14.2
# boto3==1.34.10  # Pour AWS S3

# ============================================
# Performance
# ============================================
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0

# ============================================
# Parsing & Serialization
# ============================================
xmltodict==0.13.0
pyyaml==6.0.1

# ============================================
# Chart Generation (pour rapports)
# ============================================
matplotlib==3.8.2
seaborn==0.13.1

# ============================================
# Environment Specific
# ============================================
# Décommenter selon l'environnement de déploiement
# psycopg2-binary==2.9.9  # Si PostgreSQL au lieu de MySQL
# waitress==2.1.2  # Serveur WSGI pour Windows
